update
*server sends the init message and than the init server message in
loop. The index of the queue head is fine. for some reason it
recognizes the init message and than we told him in that case to add
the server msg to the next empty slot. need to figure it out.
That seems our major problem. I believe it has something to do with the
message that was not sent.
*I changed queue head to an array instead of two ints.check this out.